Implemented #389 user confirmation of potentially dangerous actions. It works as an intermediate layer with confirmation request both in AJAX and noscript modes. Enabled it on page, forums, pfs and comments.

+++ modules/forums/tpl/forums.posts.tpl
@@ -26,7 +26,7 @@
 				<a href="{FORUMS_POSTS_ANNOUNCE_URL}" title="{PHP.L.forums_explainannounce}">{PHP.L.forums_announcement}</a> {PHP.cfg.separator}
 				<a href="{FORUMS_POSTS_PRIVATE_URL}" title="{PHP.L.forums_explainprivate}">{PHP.L.forums_private} (#)</a> {PHP.cfg.separator}
 				<a href="{FORUMS_POSTS_CLEAR_URL}" title="{PHP.L.forums_explaindefault}">{PHP.L.Default}</a> {PHP.cfg.separator}
-				<a href="{FORUMS_POSTS_DELETE_URL}" title="{PHP.L.forums_explaindelete}">{PHP.L.Delete}</a>
+				<a href="{FORUMS_POSTS_DELETE_URL}" title="{PHP.L.forums_explaindelete}" class="confirmLink">{PHP.L.Delete}</a>
 			</td>
 		</tr>
 		<tr>
